refactor(parse): extract parse
